+++ components/cluster/command/display.go
@@ -17,21 +17,15 @@ import (
 	"errors"
 	"fmt"
 	"net/url"
-	"sort"
-	"strings"
 	"time"
 
-	"github.com/fatih/color"
 	perrs "github.com/pingcap/errors"
-	"github.com/pingcap/tiup/pkg/cliutil"
 	"github.com/pingcap/tiup/pkg/cluster/api"
-	"github.com/pingcap/tiup/pkg/cluster/clusterutil"
 	operator "github.com/pingcap/tiup/pkg/cluster/operation"
 	"github.com/pingcap/tiup/pkg/cluster/spec"
 	"github.com/pingcap/tiup/pkg/cluster/task"
 	"github.com/pingcap/tiup/pkg/logger/log"
 	"github.com/pingcap/tiup/pkg/meta"
-	"github.com/pingcap/tiup/pkg/set"
 	"github.com/spf13/cobra"
 )
 
@@ -64,11 +58,9 @@ func newDisplayCmd() *cobra.Command {
 				return displayDashboardInfo(clusterName)
 			}
 
-			if err := displayClusterMeta(clusterName, &gOpt); err != nil {
-				return err
-			}
-			if err := displayClusterTopology(clusterName, &gOpt); err != nil {
-				return err
+			err = deployer.Display(clusterName, gOpt)
+			if err != nil {
+				return perrs.AddStack(err)
 			}
 
 			metadata, err := spec.ClusterMetadata(clusterName)
@@ -121,21 +113,6 @@ func displayDashboardInfo(clusterName string) error {
 	return nil
 }
 
-func displayClusterMeta(clusterName string, opt *operator.Options) error {
-	clsMeta, err := spec.ClusterMetadata(clusterName)
-	if err != nil && !errors.Is(perrs.Cause(err), meta.ErrValidate) &&
-		!errors.Is(perrs.Cause(err), spec.ErrNoTiSparkMaster) {
-		return err
-	}
-
-	cyan := color.New(color.FgCyan, color.Bold)
-
-	fmt.Printf("TiDB Cluster: %s\n", cyan.Sprint(clusterName))
-	fmt.Printf("TiDB Version: %s\n", cyan.Sprint(clsMeta.Version))
-
-	return nil
-}
-
 func destroyTombstoneIfNeed(clusterName string, metadata *spec.ClusterMeta, opt operator.Options) error {
 	topo := metadata.Topology
 
@@ -174,123 +151,4 @@ func destroyTombstoneIfNeed(clusterName string, metadata *spec.ClusterMeta, opt 
 	log.Infof("Destroy success")
 
 	return spec.SaveClusterMeta(clusterName, metadata)
-}
-
-func displayClusterTopology(clusterName string, opt *operator.Options) error {
-	metadata, err := spec.ClusterMetadata(clusterName)
-	if err != nil && !errors.Is(perrs.Cause(err), meta.ErrValidate) &&
-		!errors.Is(perrs.Cause(err), spec.ErrNoTiSparkMaster) {
-		return err
-	}
-
-	topo := metadata.Topology
-
-	clusterTable := [][]string{
-		// Header
-		{"ID", "Role", "Host", "Ports", "OS/Arch", "Status", "Data Dir", "Deploy Dir"},
-	}
-
-	ctx := task.NewContext()
-	err = ctx.SetSSHKeySet(spec.ClusterPath(clusterName, "ssh", "id_rsa"),
-		spec.ClusterPath(clusterName, "ssh", "id_rsa.pub"))
-	if err != nil {
-		return perrs.AddStack(err)
-	}
-
-	err = ctx.SetClusterSSH(topo, metadata.User, gOpt.SSHTimeout)
-	if err != nil {
-		return perrs.AddStack(err)
-	}
-
-	filterRoles := set.NewStringSet(opt.Roles...)
-	filterNodes := set.NewStringSet(opt.Nodes...)
-	pdList := topo.GetPDList()
-	for _, comp := range topo.ComponentsByStartOrder() {
-		for _, ins := range comp.Instances() {
-			// apply role filter
-			if len(filterRoles) > 0 && !filterRoles.Exist(ins.Role()) {
-				continue
-			}
-			// apply node filter
-			if len(filterNodes) > 0 && !filterNodes.Exist(ins.ID()) {
-				continue
-			}
-
-			dataDir := "-"
-			insDirs := ins.UsedDirs()
-			deployDir := insDirs[0]
-			if len(insDirs) > 1 {
-				dataDir = insDirs[1]
-			}
-
-			status := ins.Status(pdList...)
-			// Query the service status
-			if status == "-" {
-				e, found := ctx.GetExecutor(ins.GetHost())
-				if found {
-					active, _ := operator.GetServiceStatus(e, ins.ServiceName())
-					if parts := strings.Split(strings.TrimSpace(active), " "); len(parts) > 2 {
-						if parts[1] == "active" {
-							status = "Up"
-						} else {
-							status = parts[1]
-						}
-					}
-				}
-			}
-			clusterTable = append(clusterTable, []string{
-				color.CyanString(ins.ID()),
-				ins.Role(),
-				ins.GetHost(),
-				clusterutil.JoinInt(ins.UsedPorts(), "/"),
-				cliutil.OsArch(ins.OS(), ins.Arch()),
-				formatInstanceStatus(status),
-				dataDir,
-				deployDir,
-			})
-
-		}
-	}
-
-	// Sort by role,host,ports
-	sort.Slice(clusterTable[1:], func(i, j int) bool {
-		lhs, rhs := clusterTable[i+1], clusterTable[j+1]
-		// column: 1 => role, 2 => host, 3 => ports
-		for _, col := range []int{1, 2} {
-			if lhs[col] != rhs[col] {
-				return lhs[col] < rhs[col]
-			}
-		}
-		return lhs[3] < rhs[3]
-	})
-
-	cliutil.PrintTable(clusterTable, true)
-
-	return nil
-}
-
-func formatInstanceStatus(status string) string {
-	lowercaseStatus := strings.ToLower(status)
-
-	startsWith := func(prefixs ...string) bool {
-		for _, prefix := range prefixs {
-			if strings.HasPrefix(lowercaseStatus, prefix) {
-				return true
-			}
-		}
-		return false
-	}
-
-	switch {
-	case startsWith("up|l"): // up|l, up|l|ui
-		return color.HiGreenString(status)
-	case startsWith("up"):
-		return color.GreenString(status)
-	case startsWith("down", "err"): // down, down|ui
-		return color.RedString(status)
-	case startsWith("tombstone", "disconnected"), strings.Contains(status, "offline"):
-		return color.YellowString(status)
-	default:
-		return status
-	}
 }
