+++ components/dm/command/display.go
@@ -13,25 +13,15 @@
 
 package command
 
-/*
 import (
-	"errors"
-	"fmt"
-	"sort"
-	"strings"
 	"sync"
 	"time"
 
-	"github.com/fatih/color"
 	perrs "github.com/pingcap/errors"
-	"github.com/pingcap/tiup/pkg/cliutil"
+	"github.com/pingcap/tiup/components/dm/spec"
 	"github.com/pingcap/tiup/pkg/cluster/api"
 	"github.com/pingcap/tiup/pkg/cluster/clusterutil"
-	"github.com/pingcap/tiup/pkg/cluster/meta"
 	operator "github.com/pingcap/tiup/pkg/cluster/operation"
-	"github.com/pingcap/tiup/pkg/cluster/task"
-	"github.com/pingcap/tiup/pkg/set"
-	tiuputils "github.com/pingcap/tiup/pkg/utils"
 	"github.com/spf13/cobra"
 	"go.uber.org/zap"
 )
@@ -49,17 +39,18 @@ func newDisplayCmd() *cobra.Command {
 			}
 
 			clusterName = args[0]
-			if err := displayDMMeta(clusterName, &gOpt); err != nil {
-				return err
-			}
-			if err := displayClusterTopology(clusterName, &gOpt); err != nil {
-				return err
-			}
 
-			metadata, err := meta.DMMetadata(clusterName)
-			if err != nil && !errors.Is(perrs.Cause(err), meta.ValidateErr) {
+			err := deployer.Display(clusterName, gOpt)
+			if err != nil {
 				return perrs.AddStack(err)
 			}
+
+			metadata := new(spec.DMMeta)
+			err = dmspec.Metadata(clusterName, metadata)
+			if err != nil {
+				return perrs.AddStack(err)
+			}
+
 			return clearOutDatedEtcdInfo(clusterName, metadata, gOpt)
 		},
 	}
@@ -70,25 +61,7 @@ func newDisplayCmd() *cobra.Command {
 	return cmd
 }
 
-func displayDMMeta(clusterName string, opt *operator.Options) error {
-	if tiuputils.IsNotExist(meta.ClusterPath(clusterName, meta.MetaFileName)) {
-		return perrs.Errorf("cannot display non-exists cluster %s", clusterName)
-	}
-
-	clsMeta, err := meta.DMMetadata(clusterName)
-	if err != nil && !errors.Is(perrs.Cause(err), meta.ValidateErr) {
-		return err
-	}
-
-	cyan := color.New(color.FgCyan, color.Bold)
-
-	fmt.Printf("DM Cluster: %s\n", cyan.Sprint(clusterName))
-	fmt.Printf("DM Version: %s\n", cyan.Sprint(clsMeta.Version))
-
-	return nil
-}
-
-func clearOutDatedEtcdInfo(clusterName string, metadata *meta.DMMeta, opt operator.Options) error {
+func clearOutDatedEtcdInfo(clusterName string, metadata *spec.DMMeta, opt operator.Options) error {
 	topo := metadata.Topology
 
 	existedMasters := make(map[string]struct{})
@@ -154,121 +127,3 @@ func clearOutDatedEtcdInfo(clusterName string, metadata *meta.DMMeta, opt operat
 	// return any one error
 	return <-errCh
 }
-
-func displayClusterTopology(clusterName string, opt *operator.Options) error {
-	metadata, err := meta.DMMetadata(clusterName)
-	if err != nil && !errors.Is(perrs.Cause(err), meta.ValidateErr) {
-		return err
-	}
-
-	topo := metadata.Topology
-
-	clusterTable := [][]string{
-		// Header
-		{"ID", "Role", "Host", "Ports", "Status", "Data Dir", "Deploy Dir"},
-	}
-
-	ctx := task.NewContext()
-	err = ctx.SetSSHKeySet(meta.ClusterPath(clusterName, "ssh", "id_rsa"),
-		meta.ClusterPath(clusterName, "ssh", "id_rsa.pub"))
-	if err != nil {
-		return perrs.AddStack(err)
-	}
-
-	err = ctx.SetClusterSSH(topo, metadata.User, gOpt.SSHTimeout)
-	if err != nil {
-		return perrs.AddStack(err)
-	}
-
-	filterRoles := set.NewStringSet(opt.Roles...)
-	filterNodes := set.NewStringSet(opt.Nodes...)
-	masterList := topo.GetMasterList()
-	for _, comp := range topo.ComponentsByStartOrder() {
-		for _, ins := range comp.Instances() {
-			// apply role filter
-			if len(filterRoles) > 0 && !filterRoles.Exist(ins.Role()) {
-				continue
-			}
-			// apply node filter
-			if len(filterNodes) > 0 && !filterNodes.Exist(ins.ID()) {
-				continue
-			}
-
-			dataDir := "-"
-			insDirs := ins.UsedDirs()
-			deployDir := insDirs[0]
-			if len(insDirs) > 1 {
-				dataDir = insDirs[1]
-			}
-
-			status := ins.Status(masterList...)
-			// Query the service status
-			if status == "-" {
-				e, found := ctx.GetExecutor(ins.GetHost())
-				if found {
-					active, _ := operator.GetServiceStatus(e, ins.ServiceName())
-					if parts := strings.Split(strings.TrimSpace(active), " "); len(parts) > 2 {
-						if parts[1] == "active" {
-							status = "Up"
-						} else {
-							status = parts[1]
-						}
-					}
-				}
-			}
-			clusterTable = append(clusterTable, []string{
-				color.CyanString(ins.ID()),
-				ins.Role(),
-				ins.GetHost(),
-				clusterutil.JoinInt(ins.UsedPorts(), "/"),
-				formatInstanceStatus(status),
-				dataDir,
-				deployDir,
-			})
-
-		}
-	}
-
-	// Sort by role,host,ports
-	sort.Slice(clusterTable[1:], func(i, j int) bool {
-		lhs, rhs := clusterTable[i+1], clusterTable[j+1]
-		// column: 1 => role, 2 => host, 3 => ports
-		for _, col := range []int{1, 2} {
-			if lhs[col] != rhs[col] {
-				return lhs[col] < rhs[col]
-			}
-		}
-		return lhs[3] < rhs[3]
-	})
-
-	cliutil.PrintTable(clusterTable, true)
-
-	return nil
-}
-
-func formatInstanceStatus(status string) string {
-	lowercaseStatus := strings.ToLower(status)
-
-	startsWith := func(prefixs ...string) bool {
-		for _, prefix := range prefixs {
-			if strings.HasPrefix(lowercaseStatus, prefix) {
-				return true
-			}
-		}
-		return false
-	}
-
-	switch {
-	case startsWith("up|l"): // up|l, up|l|ui
-		return color.HiGreenString(status)
-	case startsWith("up"):
-		return color.GreenString(status)
-	case startsWith("down", "err"): // down, down|ui
-		return color.RedString(status)
-	case startsWith("tombstone", "disconnected"), strings.Contains(status, "offline"):
-		return color.YellowString(status)
-	default:
-		return status
-	}
-}
-*/
