+++ pkg/cluster/deploy/deploy.go
@@ -7,11 +7,14 @@ import (
 	"io"
 	"io/ioutil"
 	"os"
+	"sort"
+	"strings"
 
 	"github.com/fatih/color"
 	"github.com/joomcode/errorx"
 	perrs "github.com/pingcap/errors"
 	"github.com/pingcap/tiup/pkg/cliutil"
+	"github.com/pingcap/tiup/pkg/cluster/clusterutil"
 	operator "github.com/pingcap/tiup/pkg/cluster/operation"
 	"github.com/pingcap/tiup/pkg/cluster/spec"
 	"github.com/pingcap/tiup/pkg/cluster/task"
@@ -307,6 +310,106 @@ func (d *Deployer) Exec(clusterName string, opt ExecOptions, gOpt operator.Optio
 	return nil
 }
 
+// Display cluster meta and topology.
+func (d *Deployer) Display(clusterName string, opt operator.Options) error {
+	metadata, err := d.meta(clusterName)
+	if err != nil {
+		return perrs.AddStack(err)
+	}
+
+	topo := metadata.GetTopology()
+	base := metadata.GetBaseMeta()
+
+	// display cluster meta
+	cyan := color.New(color.FgCyan, color.Bold)
+	fmt.Printf("%s Cluster: %s\n", d.sysName, cyan.Sprint(clusterName))
+	fmt.Printf("%s Version: %s\n", d.sysName, cyan.Sprint(base.Version))
+
+	// display topology
+	clusterTable := [][]string{
+		// Header
+		{"ID", "Role", "Host", "Ports", "OS/Arch", "Status", "Data Dir", "Deploy Dir"},
+	}
+
+	ctx := task.NewContext()
+	err = ctx.SetSSHKeySet(spec.ClusterPath(clusterName, "ssh", "id_rsa"),
+		spec.ClusterPath(clusterName, "ssh", "id_rsa.pub"))
+	if err != nil {
+		return perrs.AddStack(err)
+	}
+
+	err = ctx.SetClusterSSH(topo, base.User, opt.SSHTimeout)
+	if err != nil {
+		return perrs.AddStack(err)
+	}
+
+	filterRoles := set.NewStringSet(opt.Roles...)
+	filterNodes := set.NewStringSet(opt.Nodes...)
+	pdList := topo.BaseTopo().MasterList
+	for _, comp := range topo.ComponentsByStartOrder() {
+		for _, ins := range comp.Instances() {
+			// apply role filter
+			if len(filterRoles) > 0 && !filterRoles.Exist(ins.Role()) {
+				continue
+			}
+			// apply node filter
+			if len(filterNodes) > 0 && !filterNodes.Exist(ins.ID()) {
+				continue
+			}
+
+			dataDir := "-"
+			insDirs := ins.UsedDirs()
+			deployDir := insDirs[0]
+			if len(insDirs) > 1 {
+				dataDir = insDirs[1]
+			}
+
+			status := ins.Status(pdList...)
+			// Query the service status
+			if status == "-" {
+				e, found := ctx.GetExecutor(ins.GetHost())
+				if found {
+					active, _ := operator.GetServiceStatus(e, ins.ServiceName())
+					if parts := strings.Split(strings.TrimSpace(active), " "); len(parts) > 2 {
+						if parts[1] == "active" {
+							status = "Up"
+						} else {
+							status = parts[1]
+						}
+					}
+				}
+			}
+			clusterTable = append(clusterTable, []string{
+				color.CyanString(ins.ID()),
+				ins.Role(),
+				ins.GetHost(),
+				clusterutil.JoinInt(ins.UsedPorts(), "/"),
+				cliutil.OsArch(ins.OS(), ins.Arch()),
+				formatInstanceStatus(status),
+				dataDir,
+				deployDir,
+			})
+
+		}
+	}
+
+	// Sort by role,host,ports
+	sort.Slice(clusterTable[1:], func(i, j int) bool {
+		lhs, rhs := clusterTable[i+1], clusterTable[j+1]
+		// column: 1 => role, 2 => host, 3 => ports
+		for _, col := range []int{1, 2} {
+			if lhs[col] != rhs[col] {
+				return lhs[col] < rhs[col]
+			}
+		}
+		return lhs[3] < rhs[3]
+	})
+
+	cliutil.PrintTable(clusterTable, true)
+
+	return nil
+}
+
 // EditConfig let the user edit the config.
 func (d *Deployer) EditConfig(clusterName string, skipConfirm bool) error {
 	metadata, err := d.meta(clusterName)
@@ -439,4 +542,30 @@ func editTopo(origTopo spec.Topology, data []byte, skipConfirm bool) (spec.Topol
 	}
 
 	return newTopo, nil
+}
+
+func formatInstanceStatus(status string) string {
+	lowercaseStatus := strings.ToLower(status)
+
+	startsWith := func(prefixs ...string) bool {
+		for _, prefix := range prefixs {
+			if strings.HasPrefix(lowercaseStatus, prefix) {
+				return true
+			}
+		}
+		return false
+	}
+
+	switch {
+	case startsWith("up|l"): // up|l, up|l|ui
+		return color.HiGreenString(status)
+	case startsWith("up"):
+		return color.GreenString(status)
+	case startsWith("down", "err"): // down, down|ui
+		return color.RedString(status)
+	case startsWith("tombstone", "disconnected"), strings.Contains(status, "offline"):
+		return color.YellowString(status)
+	default:
+		return status
+	}
 }
